Integrate MusicXML Parser.
Consolidate the conversion script to support both midi and musicxml.
Add conversion methods to the main Magenta API.
