+++ magenta/music/__init__.py
@@ -49,6 +49,13 @@ from magenta.music.midi_synth import synthesize
 
 from magenta.music.model import BaseModel
 
+from magenta.music.musicxml_parser import MusicXMLDocument
+from magenta.music.musicxml_parser import MusicXMLParseException
+
+from magenta.music.musicxml_reader import MusicXMLConversionError
+from magenta.music.musicxml_reader import musicxml_file_to_sequence_proto
+from magenta.music.musicxml_reader import musicxml_to_sequence_proto
+
 from magenta.music.notebook_utils import play_sequence
 
 from magenta.music.sequence_generator import BaseSequenceGenerator
