+++ magenta/tools/pip/setup.py
@@ -44,7 +44,7 @@ CONSOLE_SCRIPTS = [
     'magenta.models.melody_rnn.melody_rnn_generate',
     'magenta.models.melody_rnn.melody_rnn_train',
     'magenta.models.rl_tuner.rl_tuner_train',
-    'magenta.scripts.convert_midi_dir_to_note_sequences',
+    'magenta.scripts.convert_dir_to_note_sequences',
 ]
 
 setup(
